indexer/requirements.txt
pdfplumber
openai
python-dotenv
azure-cosmos
pillow
pdf2image
pytesseract